+++ lib/processor/while_test.go
@@ -63,6 +63,40 @@ func TestWhileWithCount(t *testing.T) {
 	}
 }
 
+func TestWhileWithContentCheck(t *testing.T) {
+	conf := NewConfig()
+	conf.Type = "while"
+	conf.While.Condition.Type = "bounds_check"
+	conf.While.Condition.BoundsCheck.MaxParts = 3
+
+	procConf := NewConfig()
+	procConf.Type = "insert_part"
+	procConf.InsertPart.Content = "foo"
+	procConf.InsertPart.Index = 0
+
+	conf.While.Processors = append(conf.While.Processors, procConf)
+
+	c, err := New(conf, nil, log.Noop(), metrics.Noop())
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	exp := [][]byte{
+		[]byte(`foo`),
+		[]byte(`foo`),
+		[]byte(`foo`),
+		[]byte(`bar`),
+	}
+
+	msg, res := c.ProcessMessage(message.New([][]byte{[]byte("bar")}))
+	if res != nil {
+		t.Error(res.Error())
+	}
+	if act := message.GetAllBytes(msg[0]); !reflect.DeepEqual(act, exp) {
+		t.Errorf("Wrong result: %s != %s", act, exp)
+	}
+}
+
 func TestWhileWithCountALO(t *testing.T) {
 	conf := NewConfig()
 	conf.Type = "while"